payment: add BUS events for all direct payments calls
Note that I have re-used the existing events for now. It may be desirable
to expose more information in the future (like the type of transaction).
Clients do have the direct payment id though, so they can retrieve the information
if needed.

This fixes https://github.com/killbill/killbill/issues/179.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>